% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add.psm.count.R
\name{add.psm.count}
\alias{add.psm.count}
\title{Add PSM count and related information to msDT}
\usage{
add.psm.count(dt, param)
}
\arguments{
\item{dt}{A msDT containing the original data.}

\item{param}{The parameter or column name used for grouping.}
}
\value{
A modified data.table with added columns representing PSM count-related information.
}
\description{
This function calculates and adds PSM counts to the msDT
}
