% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/callAnnotation.R
\name{callAnnotationDbi}
\alias{callAnnotationDbi}
\title{call Annotation using AnnotationDbi}
\usage{
callAnnotationDbi(
  prtOrganism,
  annot.keys = gene_names,
  keytype = "ENSEMBL",
  annot.cols = c("SYMBOL", "GENENAME", "ENTREZID", "REFSEQ"),
  AnnDbi.multiVals = "list"
)
}
\arguments{
\item{prtOrganism}{AnnotationDb object (genome). Currently one of human, mouse and yeast}

\item{annot.keys}{keys to select records for from the database, such as ENSEMBL ids.}

\item{keytype}{keytype that matches the keys used. For instance, if annot.keys are ensml ids, the keytype must be ENSEMBL"}

\item{annot.cols}{the columns or kinds of things that can be retrieved from the database. "SYMBOL","GENENAME","ENTREZID", "REFSEQ", ...}

\item{AnnDbi.multiVals}{What should mapIds do when there are multiple values that could be returned? look at ?AnnotationDbi::mapIds > multiVals}
}
\value{
a data.table with annot.keys and annot.cols
}
\description{
call Annotation using AnnotationDbi
}
