% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/callAnnotations.R
\name{callAnnotations}
\alias{callAnnotations}
\title{Title}
\usage{
callAnnotations(
  dat,
  feature.annotation.source,
  uniprot.input = "Protein",
  uniprot.output = "Genes",
  uniprot.extra = NULL,
  uniprot.KEY = "UNIPROTKB",
  uniprot.species.name = NULL,
  uniprot.taxId = NULL,
  n.call = 99,
  uniprot.file = "AnnotationsTable.xlsx",
  shorten.gene.names = TRUE,
  split.char = ";",
  save.annotation = TRUE,
  save.annotation.path = Tables.path,
  Data.path = Data.path
)
}
\arguments{
\item{Data.path}{}
}
\description{
Title
}
\examples{

# import UniProt.ws, vsvDialogs
}
