% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{capitalize_first_char}
\alias{capitalize_first_char}
\title{Capitalize the First Letter of Each String}
\usage{
capitalize_first_char(x)
}
\arguments{
\item{x}{A character vector for which the first letter of each element should be capitalized.}
}
\value{
A character vector with the first letter of each string capitalized.
}
\description{
This function takes a character vector and returns the vector with the first letter of each string capitalized and the remaining letters in lowercase.
}
\examples{
capitalize_first_char(c("apple", "banana", "cherry"))
# Returns: "Apple" "Banana" "Cherry"
}
