% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_input_file.R
\name{check_input_file}
\alias{check_input_file}
\title{Check and Process Input File for MS Analysis}
\usage{
check_input_file(x, sheet = 1)
}
\arguments{
\item{x}{A \code{data.table} or a character string representing the path to an input file containing MS data. If a file path is provided, the file will be read into a \code{data.table}.}

\item{sheet}{An integer specifying the sheet number to read from if the input file is an Excel file. Default is \code{1}.}
}
\value{
A list with the following elements:
\itemize{
\item \code{"MSmethod"}: The identified mass spectrometry method (\code{"TMT"} or \code{"LFQ"}).
\item \code{"data.level"}: The data level of the input data (\code{"PSM-level"}, \code{"Peptide-level"}, or \code{"Protein-level"}).
\item \code{"Intensity.col"}: The identified columns corresponding to intensity values.
\item \code{"Sample.dt"}: A \code{data.table} containing metadata extracted from the input data, such as file IDs, channels, and conditions.
\item \code{"parList"}: A list of identified relevant column names for further processing.
}
}
\description{
This function checks the input file or data table for MS analysis, determines the data level (PSM, peptide, or protein), identifies relevant columns, and processes the data to extract useful metadata.
}
\details{
The \code{check_input_file} function performs several key tasks:
\itemize{
\item Determines whether the data is at the PSM, peptide, or protein level based on the presence of specific columns.
\item Identifies the mass spectrometry method (TMT or LFQ) by examining the intensity columns.
\item Extracts relevant metadata, including sample IDs, conditions, and channels, and organizes it into a structured \code{data.table}.
\item Prepares a list of relevant column names for use in downstream analyses.
}

The function supports both direct data input as a \code{data.table} and reading data from an input file. It is designed to handle data from common MS analysis software (PD and MQ) outputs.
}
\examples{
\dontrun{
  # Check an input file for MS analysis
  input_file <- "path/to/ms_data.xlsx"
  result <- check_input_file(input_file, sheet = 1)
  print(result)
}

}
