% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{clean_column_names}
\alias{clean_column_names}
\title{Clean and Standardize Data Column Names and Values}
\usage{
clean_column_names(
  dat,
  complete.check = FALSE,
  subsymDT = data.table(symout = c("\\\\~", "\\\\+", "\\\\:", "\\\\|", "\\\\-", "\\\\(",
    "\\\\)", "\\\\/", " "), symin = c("", ".", "", "", "_", "", "", ".", ""))
)
}
\arguments{
\item{dat}{A \code{data.table} or vector containing the data to be cleaned.}

\item{complete.check}{Logical; if \code{TRUE}, the function will also clean the content within the columns of the \code{data.table}. Defaults to \code{FALSE}.}

\item{subsymDT}{A \code{data.table} specifying the symbols to be replaced or removed. It should contain two columns: \code{"symout"} (the symbol to be replaced) and \code{"symin"} (the replacement). Defaults to a table with common problematic symbols.}
}
\value{
A list with two elements:
\itemize{
\item \code{"cln.nm.dat"}: The cleaned \code{data.table} or vector.
\item \code{"ref.name.tbl"}: A reference table showing the original and cleaned column names (if \code{dat} is a \code{data.table}).
}
}
\description{
This function cleans and standardizes the column names and optionally the data within a \code{data.table} or vector by replacing or removing symbols that may interfere with design formulas. It is particularly useful in preprocessing data for statistical analysis where certain characters in column names or data values can cause issues.
}
\details{
The function ensures that certain symbols commonly used in design formulas (e.g., \code{~}, \code{+}, \code{:}, \code{|}, \code{-}, \code{(}, \verb{)}, \code{/}, and spaces) are either removed or replaced in column names and, if \code{complete.check = TRUE}, within the data itself. This avoids potential conflicts when working with formulas in statistical models.

The default \code{subsymDT} table maps symbols to replacements:
\itemize{
\item \code{"~"} -> \code{""}
\item \code{"+"} -> \code{"."}
\item \code{":"} -> \code{""}
\item \code{"|"} -> \code{""}
\item \code{"-"} -> \code{"_"}
\item \code{"("} -> \code{"._"}
\item \code{")"} -> \code{"_."}
\item \code{"/"} -> \code{"."}
\item \code{" "} -> \code{""}
}
}
\examples{
library(data.table)
dat <- data.table(`Column (1)` = c(1, 2), `Another+Column` = c(3, 4))
result <- clean_column_names(dat)
cleaned_data <- result$cln.nm.dat
reference_table <- result$ref.name.tbl

}
