% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_feature_data.R
\name{clean_feature_data}
\alias{clean_feature_data}
\title{Clean Feature-level Data}
\usage{
clean_feature_data(
  dat,
  metadata,
  filter.by.quaninfo = "auto",
  isolation.interference.cutoff = 75,
  collapse_psm_method = "TagPSM",
  filter.singleshot.proteins = "BYPEPTIDE",
  min.intensity = 0.01
)
}
\arguments{
\item{dat}{A \code{data.table} containing the feature-level data.}

\item{metadata}{A \code{data.table} containing metadata associated with the MS data.}

\item{filter.by.quaninfo}{Character. Specifies the method for filtering MS data generated by PD based on quantitative information. Options are \code{"auto"}, \code{"none"}, or a specific label. Default is \code{"auto"}.}

\item{isolation.interference.cutoff}{Numeric or NULL. The threshold for filtering based on isolation interference. Features with interference above this value are removed. Default is \code{75}.}

\item{collapse_psm_method}{Character. Method for handling multiple PSM measurements per feature. Default is \code{"TagPSM"}.}

\item{filter.singleshot.proteins}{Character. Method for filtering single-shot proteins. Options are \code{"BYPEPTIDE"} or \code{"BYFEATURE"}. Default is \code{"BYPEPTIDE"}.}

\item{min.intensity}{Numeric. Minimum intensity threshold below which intensity values are considered missing. Default is \code{0.01}.}
}
\value{
A list containing the cleaned \code{msDT}, a \code{BasicStats} table with summary statistics, and \code{Isol.Interference} data.
}
\description{
This function cleans and filters feature-level (psm-level) data based on various criteria such as contaminant removal, isolation interference, and intensity thresholds. It also handles multiple PSM measurements per feature, removes single-shot proteins, and calculates basic statistics before and after filtering.
}
\details{
The function performs the following steps:
\itemize{
\item Removes rows where protein data is missing or marked as a contaminant.
\item Filters out single-shot proteins both before and after peptide filtering.
\item Removes features with intensities below the specified minimum threshold.
\item Filters features based on isolation interference and quantitative information.
\item Consolidates multiple PSM measurements per feature.
\item Outputs basic statistics before and after the cleaning process.
}
}
\examples{
\dontrun{
  clean_data <- clean_feature_data(dat = my_data, metadata = my_metadata)
}

}
