% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/collapse_fractions.R
\name{collapse_fractions}
\alias{collapse_fractions}
\title{Collapse Fractions in Mass Spectrometry Data}
\usage{
collapse_fractions(
  dat = dat,
  group = "Pool",
  method = "MAXset",
  path = Tables.path
)
}
\arguments{
\item{dat}{A \code{data.table} containing the MS data. The table should include columns such as \code{Feature},
\code{Fraction}, \code{Condition} if label-free, \code{Channel} if TMT, and \code{Intensity}.}

\item{group}{A character string indicating the grouping variable for the analysis (e.g., "Pool"). Default is "Pool".}

\item{method}{A character string specifying the method to use for combining fractions. Options include:
\itemize{
\item \code{"MAXset"}: Select the fraction with the minimum missingness and maximum intensity across conditions/channels.
\item \code{"MAX"}: Select the fraction with the maximum intensity across conditions/channels, regardless of missingness.
\item \code{"MED"} or \code{"MEDIAN"}: Use the median intensity across fractions.
\item \code{"SUM"}: Sum the intensities across fractions.
\item \code{"SINGLE"}: Use only peptides with a single fraction and remove others.
\item \code{"MULTI"}, \code{"SEPERATE"}, \code{"TAGFRAC"}, \code{"TAGFRC"}: Tag features by fractions and pass to summarization.
}
Default is \code{"MAXset"}.}

\item{path}{A character string specifying the path to save the output summary tables. Default is \code{Tables.path}.}
}
\value{
A list with two elements:
\itemize{
\item \code{"dat.fcomb"}: A \code{data.table} containing the combined fractions data.
\item \code{"PrtPepCount_fraction"}: A \code{data.table} summarizing the protein and peptide counts before and after fraction combination.
}
}
\description{
This function combines fractions in mass spectrometry data using various methods to summarize the intensities
of features across fractions. The function supports multiple methods for summarizing the fractions, including
maximum intensity, median intensity, sum of intensities, and tagging fractions.
}
\details{
The function first identifies unique and non-unique measurements (features measured in more than one fraction per group).
It then combines the fractions based on the specified method. The function can also generate a summary of the protein and
peptide counts before and after fraction combination.
}
