% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/collapse_psm.R
\name{collapse_psm}
\alias{collapse_psm}
\title{Collapse Multiple PSM Measurements per feature}
\usage{
collapse_psm(dat, method = "IONSSCORE", by)
}
\arguments{
\item{dat}{A \code{data.table} containing the PSM data.}

\item{method}{Character. The method used to collapse multiple PSM measurements. Options include \code{"IONSSCORE"}, \code{"MEAN"}, \code{"MEDIAN"}, \code{"MAX"}, and \code{"ALL"}. Default is \code{"IONSSCORE"}.}

\item{by}{Character vector. Columns used for grouping when collapsing PSM measurements.}
}
\value{
A \code{data.table} with collapsed PSM measurements according to the specified method.
}
\description{
This function handles multiple PSM (Peptide Spectrum Match) measurements for a given feature, channel, and run by collapsing them into a single measurement according to the specified method.
}
\details{
The function offers several methods for collapsing multiple PSM measurements of a given feature:
\itemize{
\item \strong{IONSSCORE}: Selects the PSM with the highest ion score (ms method PD).
\item \strong{MEAN}: Computes the mean intensity of all PSMs.
\item \strong{MEDIAN}: Computes the median intensity of all PSMs.
\item \strong{MAX}: Selects the maximum intensity among the PSMs.
\item \strong{ALL}: Keeps all PSM measurements, tagging them for feature to protein summarization.
}
}
\examples{
\dontrun{
  collapsed_data <- collapse_psm(dat = my_data, method = "MEAN", by.col = c("Feature", "Channel", "Run"))
}

}
