% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/collapseFraction.R
\name{collapseFractions}
\alias{collapseFractions}
\title{Fraction combination}
\usage{
collapseFractions(
  dt = dt,
  group = "Pool",
  method = "MAXset",
  path = Tables.path
)
}
\arguments{
\item{path}{}
}
\description{
Fraction combination
}
