% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config_template.R
\name{config_template}
\alias{config_template}
\title{Generate a Configuration Template for MSqb2 Analysis}
\usage{
config_template()
}
\value{
The function does not return any value but sets up a configuration template with predefined parameters. This template includes various settings for filtering, normalization, summarization, fraction combination, batch effect removal, missing data imputation, annotation, contrast definition, and reformatting of input data.
}
\description{
This function generates a configuration template for MSqb2 analysis, providing a set of predefined parameters that can be customized for specific mass spectrometry data processing tasks.
}
\details{
The \code{config_template} function initializes a set of configuration parameters for MSqb2 analysis, which can be modified by the user as needed. The parameters are grouped into several categories:

\itemize{
\item \strong{Filters}: Settings for filtering data based on quantification information, isolation interference, intensity thresholds, and more.
\item \strong{Normalization}: Parameters for normalizing the data, including method selection and subset normalization.
\item \strong{Summarization}: Options for summarizing data, including the method for median polishing.
\item \strong{Fraction Combination}: Method for combining data across fractions.
\item \strong{Batch Effect Removal}: Method for removing batch effects from the data.
\item \strong{Missing Data Imputation}: Parameters for handling missing data.
\item \strong{Metadata Filtering}: Options for filtering the dataset based on metadata.
\item \strong{Annotation}: Source of feature annotation and the organism under study.
\item \strong{Contrasts}: Parameters for defining contrasts in differential expression analysis, including pairwise contrasts and complex contrast definitions.
\item \strong{Reformatting}: Options for reformatting input data, particularly for data generated from Proteome Discoverer.
}

The template is intended to be customized by users for their specific analysis needs, ensuring that all relevant parameters are set before running the MSqb2 pipeline.
}
\examples{
\dontrun{
  # Generate a configuration template
  config_template()

  # Users can modify the parameters as needed before running the analysis
}

}
