% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_qc_tables.R
\name{create_qc_tables}
\alias{create_qc_tables}
\title{Generate Quality Control Summary Tables}
\usage{
create_qc_tables(
  dat = dat,
  metadata = metadata,
  ms.software = "PD",
  save = TRUE,
  Tables.path = Tables.path
)
}
\arguments{
\item{dat}{A \code{data.table} containing the mass spectrometry data. This table should include columns such as \code{Feature}, \code{Protein},
\code{Peptide}, \code{Pool}, \code{Channel}, \code{Fraction}, and \code{Intensity}.}

\item{metadata}{A \code{data.table} containing metadata associated with the mass spectrometry data. This table may include columns such as
\code{Pool}, \code{Channel}, and \code{Fraction}.}

\item{ms.software}{A character string specifying the mass spectrometry software used, either of \code{"PD"} and \code{"MQ"}. Default is \code{"PD"} (Proteome Discoverer).}

\item{save}{Logical. If \code{TRUE}, the QC tables are saved to an Excel file. Default is \code{TRUE}.}

\item{Tables.path}{A character string specifying the directory where the output file will be saved. Default is the value of \code{Tables.path}.}
}
\value{
A list containing the QC summary tables generated during the analysis, which may include:
\itemize{
\item \code{"PSMsperFraction"}: A table showing the overlap of features across fractions.
\item \code{"MissingPerFracChnl"}: A table summarizing the missingness of data per fraction and channel.
\item \code{"extMixFrac"}: External mix fractions data (if applicable).
}
}
\description{
This function creates quality control (QC) summary tables from MS data and metadata. It calculates basic statistics,
missingness per fraction and channel, and checks for feature overlap across fractions. The results can be saved to an Excel file
and returned as a list for further analysis.
}
\details{
The \code{create_qc_tables} function generates several summary tables that are essential for quality control in mass spectrometry-based
proteomics studies:

\itemize{
\item \strong{Fractions Overlap}: Evaluates how features are distributed across multiple fractions, particularly useful for TMT or LFQ methods.
\item \strong{Missingness Analysis}: Provides a summary of missing data points per fraction and channel.
}

The function merges the MS data with the provided metadata, computes the QC statistics, and optionally saves the
results in an Excel file for easy review and further analysis.
}
