% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DE_stat_table.R
\name{DE_stat_table}
\alias{DE_stat_table}
\title{Generate result table of the Differential Expression Analysis}
\usage{
DE_stat_table(
  fit_eb,
  dat,
  featureData = NULL,
  variable = "Protein",
  Gene.name = "Genes",
  adjust = "BH",
  save = TRUE,
  path = Tables.path,
  file.name = "StatTest_output",
  ...
)
}
\arguments{
\item{fit_eb}{A fitted empirical Bayes model object (from \code{limma}).}

\item{dat}{A data.table containing additional data to merge with the statistics table.}

\item{featureData}{Optional data.table with feature data (gene names, ...) to be merged with the statistics table.}

\item{variable}{A character string specifying the variable name for proteins or features in the data. Default is "Protein".}

\item{Gene.name}{A character string specifying the column name for gene names. Default is "Genes".}

\item{adjust}{A character string specifying the method for p-value adjustment. Default is "BH".}

\item{save}{Logical, if \code{TRUE}, saves the result to a file. Default is \code{TRUE}.}

\item{path}{A character string specifying the path to save the file. Default is \code{Tables.path}.}

\item{file.name}{A character string specifying the file name for saving the output. Default is "StatTest_output".}

\item{...}{Additional arguments passed to the \code{export_DE_result} function.}
}
\value{
A data.table containing the differential expression statistics.
}
\description{
This function generates a differential expression statistics table from a fitted empirical Bayes model, merges it with additional data if provided, and optionally saves the results to a file.
}
