% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{.check.file}
\alias{.check.file}
\title{Check and Validate File Paths}
\usage{
.check.file(input, dpth = Data.path, toData = FALSE, interactive = TRUE)
}
\arguments{
\item{input}{A named list where names are file names or patterns and values are file paths. If a value is \code{NULL},
it will be ignored.}

\item{dpth}{A character string specifying the directory path to search for files. Default is \code{Data.path}.}

\item{toData}{A logical value indicating whether to copy files to the \code{dpth} directory if they are not already there. Default is \code{FALSE}.}

\item{interactive}{A logical value indicating whether the function should prompt the user interactively to copy files. Default is \code{TRUE}.}
}
\value{
A named list of files with their paths. If files were copied, their paths in the \code{dpth} directory are returned.
}
\description{
This function verifies the existence of files specified in the \code{input} list. It checks if these files are
either full paths that exist, or if they can be found in a specified directory (\code{dpth}) or the working directory.
If files are missing, the function logs an error or warning and optionally copies the files to the \code{dpth} directory.
}
\details{
The function first checks if the files in \code{input} are full paths and exist. If so, it updates the \code{output} list
and sets \code{toData} to \code{TRUE}. If not, it searches for these files in the specified \code{dpth} directory and, if necessary,
in the working directory. If files are still not found, it logs a fatal message.
If \code{toData} is \code{TRUE} and \code{interactive} is \code{TRUE}, the user will be prompted to copy missing files to the \code{dpth} directory.
}
\examples{
\dontrun{
# Example usage
input_files <- list(file1 = "path/to/file1.csv", file2 = "path/to/file2.csv")
output_files <- .check.file(input = input_files, dpth = tempdir(), toData = TRUE, interactive = TRUE)
print(output_files)
}
}
