% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{.checkBuildArgs}
\alias{.checkBuildArgs}
\title{Validate Build Arguments}
\usage{
.checkBuildArgs()
}
\description{
This function performs validation checks on several build arguments. It ensures
that the arguments conform to expected types, lengths, and values. The function
utilizes various assertion functions to validate the arguments and log assertions
using the \code{MSqb2} package's logging functions.
}
\details{
The function imports objects from the parent environment using \code{impTopEnv}
and then performs validation on the following arguments:
\itemize{
\item \code{measurements.file} - Must be a character vector.
\item \code{metadata.file} - Must be a character vector of length 1.
\item \code{measurements.file.sheet} - Either a character vector of length 1 or an integer, or NULL.
\item \code{metadata.file.sheet} - Either a character vector of length 1 or an integer, or NULL.
\item \code{ms.software} - Must be one of the specified choices ("MQ", "PD").
\item \code{analysis.name} - Character vector of length 1 or NULL.
\item \code{ms.software} - Character vector of length 1, not NULL.
\item \code{prefix} - Character vector of length 1 or NULL.
\item \code{suffix} - Character vector of length 1 or NULL.
\item \code{add.date.tag} - Logical value.
}
}
