% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{.l2w}
\alias{.l2w}
\title{Convert Long Data to Wide Format}
\usage{
.l2w(dl, col.p, row.p, val, sp = "&.&.&", asMat = FALSE, Mat.rownm = NULL, ...)
}
\arguments{
\item{dl}{A \code{data.table} in long format to be converted.}

\item{col.p}{A character vector specifying the column properties to be used in the wide format.}

\item{row.p}{A character vector specifying the row properties to be used in the wide format.}

\item{val}{A character string specifying the name of the value column in the input.}

\item{sp}{A character string used as a separator to combine multiple identifiers for row and column properties. Defaults to \code{"&.&.&"}.}

\item{asMat}{Logical; if \code{TRUE}, returns the result as a matrix. Defaults to \code{FALSE}.}

\item{Mat.rownm}{A character vector specifying the columns to be used as row names in the matrix output.}

\item{...}{Additional arguments passed to \code{dcast.data.table}, such as \code{fun.aggregate}.}
}
\value{
A wide-format \code{data.table} or matrix.
}
\description{
Converts a long-format \code{data.table} into a wide-format data table or matrix, where specific columns represent row and column properties. The function also allows combining multiple identifiers using a specified separator.
}
\details{
This function reshapes long-format data into wide format, combining row and column properties using the specified separator. It is useful for preparing data for downstream analysis, where each unique combination of row and column properties is represented as a separate column.

The function handles both \code{data.table} objects and allows the output to be returned as either a \code{data.table} or a matrix. If a matrix is returned, optional row names can be specified.
}
\examples{
# Example with a data.table input
library(data.table)
long_data <- data.table(
  Category = rep(c("A", "B", "C"), each = 3),
  Group = rep(c("X", "Y", "Z"), 3),
  Abundance = 1:9
)
wide_data <- .l2w(long_data, col.p = "Group", row.p = "Category", val = "Abundance")

# Example returning a matrix
wide_matrix <- .l2w(long_data, col.p = "Group", row.p = "Category", val = "Abundance", asMat = TRUE, Mat.rownm = "Category")

}
