% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{.loggAssert}
\alias{.loggAssert}
\title{Log and Assert Conditions}
\usage{
.loggAssert(xpr)
}
\arguments{
\item{xpr}{An expression containing the assertion(s) to evaluate. The expression should be a call to a checkmate
assertion function (e.g., \code{assertCharacter}, \code{assertLogical}) without the \code{add} parameter.}
}
\value{
The function does not return a value. It stops execution and logs an error if any assertion fails.
}
\description{
Asserts conditions and logs error messages if any assertions fail. This function captures and processes
assertion expressions, evaluates them, and logs errors using the \code{.logg} function if the assertions are not met.
}
\details{
The function captures the provided expression, modifies it to include the \code{add} parameter for collecting assertion errors,
and then evaluates the modified expression. If any assertion fails, the error messages are collected and logged using the
\code{.logg} function at the error level.
}
