% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{.match_ModelFormula_metadata}
\alias{.match_ModelFormula_metadata}
\title{Match Model Formula with Metadata}
\usage{
.match_ModelFormula_metadata(dsgn, frm)
}
\arguments{
\item{dsgn}{A data frame or data table containing the metadata design information.}

\item{frm}{A character string representing the model formula to be matched with the metadata design.}
}
\value{
The updated model formula as a character string, with any missing parameters removed if the user chooses to proceed.
}
\description{
Checks whether all parameters used in a model formula match the column names in a given metadata design table. If any parameters do not match, the user is prompted to either stop the process and correct the issue or proceed by removing the missing parameters from the model formula.
}
\details{
The function extracts all variables used in the model formula and compares them with the column names in the metadata design. If any variables are missing, a message is logged, and the user is prompted to choose whether to stop the process or remove the missing variables from the formula and continue. This function is useful in cases where the metadata design does not fully align with the model formula.
}
