% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{.w2l}
\alias{.w2l}
\title{Convert Wide Data to Long Format}
\usage{
.w2l(dw, val, col.p, row.p, sp = "&.&.&")
}
\arguments{
\item{dw}{A \code{data.table} or \code{matrix} in wide format to be converted.}

\item{val}{A character string specifying the name of the value column in the output.}

\item{col.p}{A character vector specifying the column names in the output.}

\item{row.p}{A character vector specifying the column names in the output.}

\item{sp}{A character string used as a separator to split the identifiers for row and column properties. Defaults to \code{"&.&.&"}.}
}
\value{
A \code{data.table} in long format with the specified row and column properties, and a value column.
}
\description{
Converts a wide-format data table or matrix into long format, where specific columns represent row and column properties. The function also allows splitting identifiers based on a separator and reshapes the data accordingly.
}
\details{
This function reshapes wide-format data (such as a matrix) into long format, splitting row and column properties based on a specified separator. It is useful for preparing data for downstream analysis, where each unique combination of row and column properties is represented as a separate row in the output.

The function handles both matrices and \code{data.table} objects, allowing for flexibility in input types. Row and column properties are split based on the specified separator and are expanded into new columns.
}
