% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{expandList}
\alias{expandList}
\title{Expand List into Parent Environment}
\usage{
expandList(x)
}
\arguments{
\item{x}{A named list where each element will be imported into the parent environment.}
}
\description{
This function takes a named list of objects and imports these objects into
the parent environment of the function call. The objects in the list are
retrieved using their names and placed into the environment one level up
from the current environment.
}
\details{
The function uses \code{mget} to retrieve objects from the environment
specified by the list \code{x} and \code{list2env} to export them to the parent environment.
This is useful when you want to make multiple objects from a list available
in the environment where the function was called.
}
\examples{
# Example usage:
a <- 1
b <- 2
my_list <- list(a = a, b = b)
{
  c <- 3
  expandList(my_list)
  # Now `a` and `b` are available in this environment
  print(a) # Should print 1
  print(b) # Should print 2
}

}
