% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{expTopEnv}
\alias{expTopEnv}
\title{Export Objects to Parent Environment}
\usage{
expTopEnv()
}
\description{
This function exports all objects from the current environment to the parent
environment. It retrieves the objects from the current environment and places
them into the environment that is one level up from the current environment.
}
\details{
The function uses \code{mget} to get the objects from the current environment
and \code{list2env} to export them to the parent environment. This is useful when you
need to make objects available to a higher-level scope.
}
\examples{
# Example usage:
{
  a <- 1
  b <- 2
  # a and b are in the current environment
  expTopEnv()
  # a and b should now be available in the parent environment
  print(parent.frame()$a) # Should print 1
  print(parent.frame()$b) # Should print 2
}

}
