% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{fact2char}
\alias{fact2char}
\title{Convert Factor Columns to Character Columns in a Data Table}
\usage{
fact2char(dat)
}
\arguments{
\item{dat}{A \code{data.table} object in which factor columns need to be converted to character columns.}
}
\value{
A \code{data.table} object with factor columns converted to character columns.
}
\description{
Converts all factor columns in a data.table to character columns.
}
\details{
The function identifies all columns in the \code{data.table} that are of type factor and converts them to character columns.
If no factor columns are present, the original \code{data.table} is returned unchanged.
}
\examples{
library(data.table)
dat <- data.table(a = factor(c("x", "y", "z")), b = 1:3)
dat <- fact2char(dat)
str(dat) # 'a' column is now character

}
