% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filterFeatLevelData.R
\name{filterFeatLevelData}
\alias{filterFeatLevelData}
\title{Title}
\usage{
filterFeatLevelData(
  msd,
  metadata,
  filter.by.quaninfo = "auto",
  isolation.interference.cutoff = 75,
  multi.features.method = "TagPSM",
  filter.singleshot.proteins = "BYPEPTIDE",
  min.intensity = 0.01
)
}
\arguments{
\item{min.intensity}{}
}
\description{
Title
}
