% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{impTopEnv}
\alias{impTopEnv}
\title{Import Objects from Parent Environment}
\usage{
impTopEnv()
}
\description{
This function imports all objects from the parent environment of the caller
into the current environment. It retrieves the objects from the environment
that is two levels up from the current environment and places them into the
environment that is one level up.
}
\details{
The function uses \code{mget} to get the objects from the parent environment
and \code{list2env} to import them into the current environment. It is useful when
you need to bring objects from a higher level environment into the current scope.
}
\examples{
# Example usage:
a <- 1
b <- 2
{
  c <- 3
  impTopEnv()
  # Now `a` and `b` are available in this environment
  print(a) # Should print 1
  print(b) # Should print 2
}

}
