% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/impute_missing_values.R
\name{impute_missing_values}
\alias{impute_missing_values}
\title{Impute Missing Values in Data}
\usage{
impute_missing_values(
  dat,
  val = "Abundance",
  col.p = "Filename",
  row.p = "Feature",
  na.imputation.method = "v2-mnar",
  hybrid.mar = "KNN",
  hybrid.mnar = "QRILC",
  n_knn = 15,
  tune.sigma = 0.5,
  q = 0.01,
  ...
)
}
\arguments{
\item{dat}{A \code{data.table} containing the data to be imputed. The table should include columns corresponding to features, samples,
and values to be imputed.}

\item{val}{A character string specifying the column name that contains the values to be imputed. Default is \code{"Abundance"}.}

\item{col.p}{A character string specifying the column name(s) that either contains unique sample names (e.g., \code{"Filename"}), or a combination of which can be used to uniqule address each sample (e.g., \code{c("Pool", "Channel")} for MS method TMT). Default is \code{"Filename"}.}

\item{row.p}{A character string specifying the column name that contains the feature identifiers. Default is \code{"Feature"}.}

\item{na.imputation.method}{A character string specifying the method to use for imputation. Options include:
\itemize{
\item \code{"V2"}: method v2 of the msImpute package
\item \code{"V2-MNAR"}: method v2-mnar of the msImpute method specifically for MNAR data.
\item \code{"QRILC"}: Quantile Regression Imputation of Left-Censored data.
\item \code{"KNN"}: K-nearest neighbors.
\item \code{"MLE"}: Maximum Likelihood Estimation.
\item \code{"MINPROB"}: Impute using the minimal probability.
\item \code{"MINDET"}: Impute using the minimal detection limit.
\item \code{"ZERO"}: Impute with zeros.
\item \code{"HYBRID"}: Hybrid method combining MAR and MNAR strategies.
}
Default is \code{"v2-mnar"}.}

\item{hybrid.mar}{A character string specifying the method to use for MAR imputation when using the hybrid approach. Default is \code{"KNN"}.}

\item{hybrid.mnar}{A character string specifying the method to use for MNAR imputation when using the hybrid approach. Default is \code{"QRILC"}.}

\item{n_knn}{An integer specifying the number of nearest neighbors to use for KNN imputation. Default is \code{15}.}

\item{tune.sigma}{A numeric value specifying the tuning parameter for the sigma in QRILC and MinProb methods. Default is \code{0.5}.}

\item{q}{A numeric value specifying the quantile for MinDet and MinProb methods. Default is \code{0.01}.}

\item{...}{Additional arguments passed to the imputation functions.}
}
\value{
A \code{data.table} containing the original data with missing values imputed.
}
\description{
This function imputes missing values in a data set using various imputation methods. It supports a range of imputation techniques
including K-nearest neighbors (KNN), maximum likelihood estimation (MLE), quantile regression imputation for left-censored data (QRILC),
and hybrid methods combining different strategies for missing at random (MAR) and missing not at random (MNAR) data.
}
\details{
The \code{impute_missing_values} function offers several methods for imputing missing data, depending on the nature of the missingness and the experimental design. For in-depth details please refere to vignette of the packages msImpute and imputeLCMD.

\itemize{
\item \code{"v2"}: This method, provided by the \code{msImpute} package, is designed for missing at random (MAR) data. It is particularly suitable for DIA experiments where missing values are assumed to be random.
\item \code{"v2-mnar"}: This method extends \code{"v2"} to handle missing not at random (MNAR) data, making it suitable for cases where missing values are more likely for low-abundance features. This is often the case in DDA (like TMT) experiments, where the absence of data might be systematically related to the intensity of the signal.
}
}
