% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{int2fact}
\alias{int2fact}
\title{integer to factor converter}
\usage{
int2fact(dt)
}
\arguments{
\item{dt}{Input data of class data.table or data.frame.}
}
\value{
The input data \code{dt} with columns of class integer converted to factor.
}
\description{
Converts data.frame or data.table columns of class integer to factor.
}
\examples{
DT <- data.table(x = c(1., 2.), y = c(1L, 2L), z = c("1", "2"))
# DT <- data.frame(x = c(1.,2.), y = c(1L,2L), z = c("1","2"))
sapply(DT, class)
int2fact(DT)
sapply(DT, class)
}
\seealso{
\code{\link{MSqb2::char2fact}}
}
