% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DE_stat_func.R
\name{DE_stat_func}
\alias{DE_stat_func}
\title{Title}
\usage{
DE_stat_func(
  data,
  sample = "SampleID",
  variable = "Protein",
  value = "Abundance",
  model.formula = "~ 0 + Condition",
  sva.obj = NULL,
  techrep = NULL,
  limma.trend = TRUE,
  limma.robust = TRUE,
  limma.winsor.tail.p = FALSE,
  time.course = FALSE,
  spline.df = 4,
  removeTotalMissing = TRUE,
  pairwise.contrasts = TRUE,
  pairwise.denominator = c("wt", "wildtype", "wild-type", "cntrl", "control", "ctl",
    "untreated"),
  manual.contrasts = NULL,
  choose.contrasts = FALSE,
  make.complex.contrasts = FALSE,
  blocking.parameter = NULL,
  reverse.contrasts = FALSE,
  showWarnings = FALSE,
  prefix = NULL,
  suffix = NULL,
  rmSingleShotProteins = TRUE,
  add.vs.inContrast = TRUE,
  path = getwd(),
  ...
)
}
\arguments{
\item{path}{}
}
\description{
Title
}
