% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/median_normalisation.R
\name{median_normalisation}
\alias{median_normalisation}
\title{Median Normalization of Data}
\usage{
median_normalisation(
  dat,
  col.p,
  row.p,
  val,
  log2transform = TRUE,
  normalize.bySubset = NULL,
  normalize.bySubset.matchTerm = NULL,
  normalize.bySubset.matchTerm.exactMatch = FALSE
)
}
\arguments{
\item{dat}{A \code{data.table} containing the data to be normalized. The table should include columns for the features, samples, and values to be normalized.}

\item{col.p}{A character string specifying the column name that contains the sample identifiers (e.g., \code{"SampleID"}).}

\item{row.p}{A character string specifying the column name that contains the feature identifiers (e.g., \code{"Protein"}).}

\item{val}{A character string specifying the column name that contains the values to be normalized (e.g., \code{"Abundance"}).}

\item{log2transform}{Logical. If \code{TRUE}, log2 transformation is applied to the data before normalization. Default is \code{TRUE}.}

\item{normalize.bySubset}{A character string specifying the column name to be used for subsetting the data before normalization. Default is \code{NULL}.}

\item{normalize.bySubset.matchTerm}{A character string specifying the term to match within the subset column for normalization. Default is \code{NULL}.}

\item{normalize.bySubset.matchTerm.exactMatch}{Logical. If \code{TRUE}, requires an exact match for the \code{normalize.bySubset.matchTerm}. Default is \code{FALSE}.}
}
\value{
A \code{data.table} containing the normalized data, with the same structure as the input \code{dat}.
}
\description{
This function performs median normalization on a dataset. Optionally, it can log2 transform the data before normalization
and perform normalization based on a subset of the data specified by the user.
}
\details{
The \code{median_normalisation} function applies median normalization to the data, which adjusts the values in each sample such that
the median value is consistent across samples. This is particularly useful in MS or other omics data where
global shifts in signal intensities between samples need to be corrected.

The function allows for:
\itemize{
\item \strong{Log2 Transformation}: Optional log2 transformation of the data prior to normalization.
\item \strong{Subset Normalization}: Normalization based on a subset of the data, defined by the \code{normalize.bySubset} parameter.
}
If \code{normalize.bySubset} is provided, the function calculates the median based on the specified subset, otherwise, it normalizes
the data using the entire dataset. The function also supports partial or exact matching for subsetting.
}
\examples{
\dontrun{
  data <- data.table(
    SampleID = c("S1", "S2", "S3", "S4"),
    Protein = c("P1", "P2", "P1", "P2"),
    Abundance = c(1.2, 2.3, 3.1, 4.2)
  )

  normalized_data <- median_normalisation(data, col.p = "SampleID", row.p = "Protein", val = "Abundance")
}

}
