% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MEDnorm.R
\name{MEDnorm}
\alias{MEDnorm}
\title{Title}
\usage{
MEDnorm(
  dt,
  col.p,
  row.p,
  val,
  log2transform = TRUE,
  normalize.bySubset = NULL,
  normalize.bySubset.matchTerm = NULL,
  normalize.bySubset.matchTerm.exactMatch = FALSE
)
}
\arguments{
\item{normalize.bySubset.matchTerm.exactMatch}{}
}
\description{
Title
}
