% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/merge_ms_metadata.R
\name{merge_ms_metadata}
\alias{merge_ms_metadata}
\title{Merge Metadata with Mass Spectrometry Data}
\usage{
merge_ms_metadata(msDT, metadata, extMixFrac, int.nm)
}
\arguments{
\item{msDT}{A \code{data.table} containing the mass spectrometry data. This table should include columns such as \code{Pool}, \code{Channel}, and \code{Protein}.}

\item{metadata}{A \code{data.table} containing the metadata to be merged with \code{msDT}. This table should include columns such as \code{Pool} and \code{Channel}.}

\item{extMixFrac}{A \code{data.table} containing additional information for external mix fractions, used if \code{Pool} information is missing in the metadata.}

\item{int.nm}{A character string specifying the internal name used for processing. (This parameter is referenced in the function but not used within the provided code.)}
}
\value{
A list containing two elements:
\itemize{
\item \code{"msDT"}: The merged mass spectrometry data as a \code{data.table}.
\item \code{"metadata"}: The merged metadata as a \code{data.table}.
}
}
\description{
This function merges metadata with mass spectrometry data, ensuring consistency between the pools and channels in both datasets.
It performs checks to validate the presence and consistency of \code{Pool} and \code{Channel} information between the datasets, and adjusts
them as necessary to match.
}
\details{
The \code{merge_ms_metadata} function ensures that the \code{Pool} and \code{Channel} columns in the mass spectrometry data (\code{msDT}) and the metadata are consistent.
It checks for the presence of \code{Pool} information in both datasets, adjusts levels if necessary, and merges the metadata into the mass spectrometry data.

Specific steps include:
\itemize{
\item Verifying and correcting \code{Pool} levels between \code{msDT} and \code{metadata}.
\item Handling cases where \code{Pool} information is missing in either \code{msDT} or \code{metadata}.
\item Ensuring consistency of \code{Channel} information between the datasets.
\item Merging the metadata into \code{msDT} with consideration for potential inconsistencies.
}
}
