% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mergeMetadata.R
\name{mergeMetadata}
\alias{mergeMetadata}
\title{Title}
\usage{
mergeMetadata(msDT, metadata, extMixFrac, int.nm)
}
\arguments{
\item{int.nm}{}
}
\description{
Title
}
\examples{
dt <- data.table(ch = c("a", "b", "c"), Pool = rep(c("pa", "pb"), each = 3))
exp <- data.table(ch = c("a", "b", "c"), Pool = rep(c("p1", "p2", "p3"), each = 3))
mergeMetadata(dt, exp, "ch")
}
