% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mkQCtables.R
\name{mkQCtables}
\alias{mkQCtables}
\title{Title}
\usage{
mkQCtables(
  msDT = msDT,
  metadata = metadata,
  ms.software = "PD",
  save2file = TRUE,
  Tables.path = Tables.path
)
}
\arguments{
\item{Tables.path}{}
}
\description{
Title
}
