% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/msqb_build.R
\name{msqb_build}
\alias{msqb_build}
\title{Construct the skeleton of the analysis}
\usage{
msqb_build(
  measurements.file,
  metadata.file,
  measurements.file.sheet = 1,
  metadata.file.sheet = 1,
  ms.software,
  analysis.name = NULL,
  prefix = NULL,
  suffix = NULL,
  add.date.tag = TRUE,
  interactive = TRUE
)
}
\arguments{
\item{measurements.file}{character string or vector, naming the file(s) containing the MS measurements. The input file(s) must either be the PSM-level file(s) from ProteomDiscoverer or evidence file from MaxQuant. If more than one filename provided, the files will be merged to construct a single dataset.}

\item{metadata.file}{character string, naming the excel file containing the experimental design.}

\item{measurements.file.sheet}{either a character string (the name of a sheet), or an integer (the position of a sheet). Only relevant if measurements.file is from xlsx family. By default the first sheet will be read.}

\item{metadata.file.sheet}{either a character string (the name of a sheet), or an integer (the position of the sheet). By default the first sheet will be read.}

\item{ms.software}{character, naming the software used to generate ms data. Options are 'MQ' for MaxQuant and 'PD' for Proteome Discoverer.}

\item{analysis.name}{character string. If provided, a directory with the provided name will be created under the Analysis directory. The output of the workflow will then be saved in this directory.}

\item{prefix}{character string to be added at the beginning of the directory names.}

\item{suffix}{character string to be added at the end of the directory names.}

\item{add.date.tag}{logical. If TRUE, the date of the analysis will be added to the Analysis directory. Default is TRUE.}

\item{interactive}{logical.}
}
\value{
A list containing the paths and input parameters.
}
\description{
The function checks the input files and construct a folder structure to store output of the workflow.
By default, the current directory from which the function is called will be considered as the project
path and the basename of the path will be considered as the project name.
Three main directories will be created under the project path:
'Data', 'Analysis' and 'Scripts'. The output from each round of the
analysis will be saved under 'Analysis' in a separate directory named by the argument 'analysis.name'.
If 'analysis.name' is NULL, the output files will be stored directly in the 'Analysis' directory.
A log file will be generated for each call of the function 'msqb_build'. The log files have a date
and time tag and will be saved in the '_log_files' sub-directory under 'Analysis' directory.
}
