% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/msqb_viz.R
\name{msqb_viz}
\alias{msqb_viz}
\title{Title}
\usage{
msqb_viz(
  msqb_res,
  config.para.viz = NULL,
  config.viz.file = NULL,
  QC = TRUE,
  FitResidual = TRUE,
  PCA = TRUE,
  Volcano = TRUE,
  Heatmaps = TRUE,
  ProfilePlots = TRUE,
  Venndiagram = TRUE,
  LFCbars = FALSE,
  ...
)
}
\arguments{
\item{config.viz.file}{}
}
\description{
Title
}
\examples{
RColorBrewer
}
