% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalise_data.R
\name{normalise_data}
\alias{normalise_data}
\title{Normalize MS Data}
\usage{
normalise_data(method = "VSN", path = Tables.path, ...)
}
\arguments{
\item{method}{Character. The normalization method to use. Options are \code{"vsn"}, \code{"median"}, or \code{"quantile"}. Default is \code{"VSN"}.}

\item{path}{Character. The path where the output Excel file with normalized and non-normalized data will be saved. Default is \code{Tables.path}.}

\item{...}{Additional arguments passed to the normalization functions, including the column and row identifiers (\code{col.p}, \code{row.p}), the intensity column (\code{val}), and other specific parameters for each normalization method.}
}
\value{
A \code{data.table} containing the normalized data, with the intensity values replaced by the normalized abundance values.
}
\description{
This function normalizes mass spectrometry data using one of three available methods: Variance Stabilizing Normalization (VSN), median normalization, or quantile normalization. The function can also handle normalization by specific data subsets.
}
\details{
\itemize{
\item \strong{vsn}: Variance Stabilizing Normalization is applied, with an option to normalize by a subset of the data.
\item \strong{median}: Median normalization can be applied, with log2 transformation by default.
\item \strong{quantile}: Quantile normalization is performed, transforming the intensity distributions of all samples to be identical.
}

The function also generates an Excel file containing both the normalized and non-normalized peptide-level data.
}
\examples{
\dontrun{
normalized_data <- normalise_data(
  method = "VSN",
  dat = my_data,
  val = "Intensity",
  path = "output_directory"
)
}

}
