% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NormalizeData.R
\name{NormalizeData}
\alias{NormalizeData}
\title{Title}
\usage{
NormalizeData(method = "VSN", path = Tables.path, ...)
}
\arguments{
\item{...}{}
}
\description{
Title
}
\examples{
# import preprocessCore
}
