% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{order.num}
\alias{order.num}
\title{Order a \code{data.table} by Alphanumeric Column}
\usage{
order.num(dat, x)
}
\arguments{
\item{dat}{A \code{data.table} containing the data to be ordered.}

\item{x}{A string specifying the column name to be ordered. The column can be either character or factor type.}
}
\value{
The input \code{data.table} ordered by the specified column.
}
\description{
This function orders a \code{data.table} based on alphanumeric sorting of a specified column. The sorting is done by extracting numeric components from the column values while preserving the overall alphanumeric order. It is particularly useful for cases where numeric sequences are embedded within text.
}
\details{
The function extracts numeric values from the specified column and orders the \code{data.table} based on these numbers. It handles mixed alphanumeric values by focusing on the numeric part while maintaining the overall alphanumeric order. If the column is a factor, the levels are also updated to reflect the new order.

The underlying idea is based on a solution from \href{https://stackoverflow.com/questions/49084625/sort-strings-by-numbers-inside-them}{Stack Overflow}.
}
\examples{
library(data.table)
dat <- data.table(id = c("item1", "item10", "item2", "item20"))
ordered_dt <- order.num(dat, "id")
print(ordered_dt)

}
