% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Pep2Prot.R
\name{Pep2Prot}
\alias{Pep2Prot}
\title{Title}
\usage{
Pep2Prot(
  dw,
  var = "Abundance",
  groupby = "Protein",
  techrep = techrep,
  medianpolish.byPool = FALSE,
  block = "Pool",
  summarization.method = "tmp",
  medianpolish.method = "column.eff",
  rm.total.missingness = TRUE,
  Tables.path = Tables.path
)
}
\arguments{
\item{Tables.path}{}
}
\description{
Title
}
