% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{read.file}
\alias{read.file}
\title{Read Data from File}
\usage{
read.file(file, sheet = 1)
}
\arguments{
\item{file}{A character string specifying the file path to read. The file can be an Excel file (with \code{.xlsx} or \code{.xls} extension)
or a tab-delimited text file.}

\item{sheet}{An integer specifying the sheet number to read from the Excel file. Default is \code{1}.}
}
\value{
A \code{data.table} object containing the data read from the file.
}
\description{
Reads data from a file, supporting both Excel files (\code{.xlsx}, \code{.xls}) and tab-delimited text files.
The function converts the data into a \code{data.table} object and applies a conversion to factors if necessary.
}
\details{
The function checks the file extension to determine whether it should use \code{readxl::read_excel} for Excel files or
\code{data.table::fread} for text files. After reading the data, the function converts it to a \code{data.table} and applies
\code{char2fact} to ensure that character columns are converted to factors.
}
\examples{
\dontrun{
# Example usage with a tab-delimited file
data <- read.file("path/to/datafile.txt")

# Example usage with an Excel file
data <- read.file("path/to/datafile.xlsx", sheet = 2)
}

}
