% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reformat_PD_data.R
\name{reformat_PD_data}
\alias{reformat_PD_data}
\title{Reformat PSM Data from Proteome Discoverer}
\usage{
reformat_PD_data(
  feat.dt = feat.dt,
  metadata = metadata,
  protein.column = "Master Protein Accessions",
  peptide.seq.column = "Sequence",
  peptide.annot.seq.column = "Annotated Sequence",
  charge.column = "Charge",
  FileID.column = "File ID",
  ion.score.column = "Score",
  quaninfo.column = "Quan Info",
  isolation.interference.column = "Isolation Interference [\%]",
  desc.column = "Master Protein Descriptions",
  contaminant.column = "Contaminant",
  marked.as.column = "Marked as",
  abundance.columns = "Abundance",
  ms.software = "PD",
  subsymDT = data.table(symout = c("\\\\~", "\\\\+", "\\\\:", "\\\\|", "\\\\-", "\\\\(",
    "\\\\)", "\\\\/", " "), symin = c("_", ".p", ".", ".", "_", ".", ".", ".", "."))
)
}
\arguments{
\item{feat.dt}{A data.table containing PSM-level data from Proteome Discoverer.}

\item{metadata}{A data.table containing metadata for the samples.}

\item{protein.column}{Name of the column containing protein accession identifiers. Default is "Master Protein Accessions".}

\item{peptide.seq.column}{Name of the column containing peptide sequences. Default is "Sequence".}

\item{peptide.annot.seq.column}{Name of the column containing annotated peptide sequences. Default is "Annotated Sequence".}

\item{charge.column}{Name of the column containing charge states. Default is "Charge".}

\item{FileID.column}{Name of the column containing file IDs. Default is "File ID".}

\item{ion.score.column}{Name of the column containing ion scores. Default is "Score".}

\item{quaninfo.column}{Name of the column containing quantitation information. Default is "Quan Info".}

\item{isolation.interference.column}{Name of the column containing isolation interference values. Default is "Isolation Interference \link{\%}".}

\item{desc.column}{Name of the column containing protein descriptions. Default is "Master Protein Descriptions".}

\item{contaminant.column}{Name of the column indicating contaminants. Default is "Contaminant".}

\item{marked.as.column}{Name of the column indicating marked entries. Default is "Marked as".}

\item{abundance.columns}{Prefix for the columns containing abundance or intensity values. Default is "Abundance".}

\item{ms.software}{Character string indicating the mass spectrometry software used. Default is "PD".}

\item{subsymDT}{A data.table used to replace special characters in column names. Default is a predefined data.table.}
}
\value{
A list containing reformatted PSM data, metadata, and annotation data.
}
\description{
This function reformats PSM-level data specifically generated from the Proteome Discoverer software, merging it with metadata and performing necessary clean-up steps for downstream analysis.
}
\examples{
\dontrun{
# Example usage of reformat_PD_data function for data from Proteome Discoverer
reformatted_data <- reformat_PD_data(feat.dt = my_data, metadata = my_metadata)
}
}
