% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reformatRAW.R
\name{reformatRAW}
\alias{reformatRAW}
\title{Title}
\usage{
reformatRAW(
  feat.dt = feat.dt,
  metadata = metadata,
  protein.column = "Master Protein Accessions",
  peptide.seq.column = "Sequence",
  peptide.annot.seq.column = "Annotated Sequence",
  charge.column = "Charge",
  FileID.column = "File ID",
  ion.score.column = "Score",
  quaninfo.column = "Quan Info",
  isolation.interference.column = "Isolation Interference [\%]",
  desc.column = "Master Protein Descriptions",
  contaminant.column = "Contaminant",
  marked.as.column = "Marked as",
  abundance.columns = "Abundance",
  reference.channel = "reference",
  ms.software = "PD",
  subsymDT = data.table(symout = c("\\\\~", "\\\\+", "\\\\:", "\\\\|", "\\\\-", "\\\\(",
    "\\\\)", "\\\\/", " "), symin = c("_", ".p", ".", ".", "_", ".", ".", ".", "."))
)
}
\arguments{
\item{subsymDT}{}
}
\description{
Title
}
