% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remove_batch_effect.R
\name{remove_batch_effect}
\alias{remove_batch_effect}
\title{Remove Batch Effects from Data}
\usage{
remove_batch_effect(
  dat,
  metadt,
  val = "Abundance",
  batch.corr.method = "limma",
  row.p = "auto",
  col.p = "auto",
  model.formula,
  sva.n.sv = NULL
)
}
\arguments{
\item{dat}{A data.table containing the data to be corrected for batch effects.}

\item{metadt}{A data.table containing metadata corresponding to the data.}

\item{val}{A character string specifying the column name containing the values to be corrected. Default is "Abundance".}

\item{batch.corr.method}{A character string specifying the batch correction method to be used. Options are "limma" or "sva". Default is "limma".}

\item{row.p}{A character vector specifying the row identifiers in the data. Default is "auto".}

\item{col.p}{A character vector specifying the column identifiers in the data. Default is "auto".}

\item{model.formula}{A formula specifying the model for the batch correction.}

\item{sva.n.sv}{An integer specifying the number of surrogate variables to estimate if using the \code{sva} method. Default is \code{NULL}.}
}
\value{
A data.table with batch effects removed. If using \code{sva}, it returns a list containing the cleaned data and the \code{sva} object.
}
\description{
This function removes batch effects from a dataset using the methods implimented in the packages \code{limma}, \code{sva}, or others. It returns the cleaned data with batch effects removed.
}
\examples{
\dontrun{
# Example usage of remove_batch_effect function
cleaned_data <- remove_batch_effect(dat, metadt, batch.corr.method = "limma")
}
}
