% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{str_sort_numeric}
\alias{str_sort_numeric}
\title{Sort a character vector with numeric order}
\usage{
str_sort_numeric(x, decreasing = FALSE)
}
\arguments{
\item{x}{A character vector to be sorted.}

\item{decreasing}{Logical; if \code{TRUE}, the sort will be in decreasing order.}
}
\value{
A character vector sorted based on numeric values within the strings.
}
\description{
This function sorts a character vector in a way that numbers within the strings are considered as numeric values during sorting.
}
\examples{
str_sort_numeric(c("file2.txt", "file10.txt", "file1.txt"))
# Returns: "file1.txt" "file2.txt" "file10.txt"
}
