% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{subTopVar}
\alias{subTopVar}
\title{Subsets Top proteins/genes Based on Variance}
\usage{
subTopVar(
  dt,
  row.p,
  col.p,
  val,
  topNperc,
  min500 = TRUE,
  topN = NULL,
  complete.rows,
  sep.col = "&.&.&",
  returnLongFormat = FALSE
)
}
\arguments{
\item{dt}{Data table to be processed.}

\item{row.p}{Column name in \code{dt} representing the rows.}

\item{col.p}{Column name in \code{dt} representing the columns.}

\item{val}{Column name in \code{dt} representing the values.}

\item{topNperc}{Percentage of top variant rows to return. When specified, \code{topN} is ignored.}

\item{min500}{Logical; if TRUE, a minimum of 500 rows are returned.}

\item{topN}{Integer; the number of top variant rows to return. When specified, \code{topNperc} is ignored.}

\item{complete.rows}{Logical; if TRUE, rows with missing values are omitted.}

\item{sep.col}{String; separator used in the matrix conversion.}

\item{returnLongFormat}{Logical; if TRUE, returns the data in long format.}
}
\value{
A subset of \code{dt} with the top variable rows based on variance.
}
\description{
This function subsets the input data table based on the highest row-wise variance.
It can return either a specific number of top rows (\code{topN}) or a percentage
(\code{topNperc}) of the rows with the highest variance. Additional options allow
control over the minimum number of rows returned and the format of the output.
The combination of \code{row.p}, \code{col.p} and \code{val} will be used to reformat data via
function \code{dcast} from \code{data.table} package.
}
\examples{
# Example usage of subTopVar function
data(iris)
dt <- as.data.table(iris)
topVarData <- subTopVar(dt, "Species", "Sepal.Length", "Sepal.Width",
                        topNperc = 20, complete.rows = TRUE,
                        returnLongFormat = FALSE)
head(topVarData)

}
