% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vsn_normalisation.R
\name{vsn_normalisation}
\alias{vsn_normalisation}
\title{VSN Normalization}
\usage{
vsn_normalisation(
  dat,
  col.p,
  row.p,
  val,
  calib = "affine",
  normalize.bySubset = NULL,
  normalize.bySubset.matchTerm = NULL,
  normalize.bySubset.matchTerm.exactMatch = FALSE,
  sp = "_.._"
)
}
\arguments{
\item{dat}{A \code{data.table} containing the data to be normalized.}

\item{col.p}{Character vector. Column names used for grouping the data (e.g., sample or condition).}

\item{row.p}{Character vector. Row identifiers, typically the feature or protein IDs.}

\item{val}{Character. The column name representing the intensity or abundance values to be normalized.}

\item{calib}{Character. Calibration method to use, either \code{"affine"} (default) or \code{"none"}. If \code{"none"}, it assumes the data is already normalized.}

\item{normalize.bySubset}{Character or NULL. Column name used to define a subset for normalization. If \code{NULL}, the entire dataset is normalized.}

\item{normalize.bySubset.matchTerm}{Character or NULL. The term used to match entries within the subset defined by \code{normalize.bySubset}. If \code{NULL}, the whole dataset is used.}

\item{normalize.bySubset.matchTerm.exactMatch}{Logical. If \code{TRUE}, only exact matches of \code{normalize.bySubset.matchTerm} are used for subsetting. Default is \code{FALSE}.}

\item{sp}{Character. Separator used for pasting together row identifiers when converting data between wide and long formats. Default is \code{"_.._"}.}
}
\value{
A \code{data.table} with normalized intensity values.
}
\description{
This function applies Variance Stabilizing Normalization (VSN) to mass spectrometry data, with options to normalize based on subsets of the data or the entire dataset.
}
\details{
This function uses the \code{vsn2} function from the \code{vsn} package to perform normalization. The normalization can be applied to the entire dataset or to a specific subset, which can be useful if different subsets require separate normalization.
\itemize{
\item \strong{Subset Normalization:} If \code{normalize.bySubset} is specified, only the subset matching \code{normalize.bySubset.matchTerm} is used for normalization.
\item \strong{Calibration:} The \code{calib} parameter allows for affine calibration or no calibration (\code{"none"}).
}
}
\examples{
\dontrun{
  normalized_data <- vsn_normalisation(dat = my_data, col.p = "SampleID", row.p = "ProteinID", val = "Intensity")
}

}
