% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{char2fact}
\alias{char2fact}
\title{character to factor converter}
\usage{
char2fact(dat)
}
\arguments{
\item{dat}{Input data of class data.table or data.frame.}
}
\value{
The input data \code{dat} with columns of class character converted to factor.
}
\description{
Converts data.frame or data.table columns of class character to factor.
}
\examples{
DT <- data.table(x = c(1., 2., 3.), y = c(4L, 5L, 6L), z = c("a", "b", "NA"))
sapply(DT, class)
char2fact(DT)
sapply(DT, class)
}
\seealso{
\code{\link{int2fact}}
}
