% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_QC_tables.R
\name{create_QC_tables}
\alias{create_QC_tables}
\title{Title}
\usage{
create_QC_tables(
  msDT = msDT,
  metadata = metadata,
  ms.software = "PD",
  save2file = TRUE,
  Tables.path = Tables.path
)
}
\arguments{
\item{Tables.path}{}
}
\description{
Title
}
